+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Define `WP_ENVIRONMENT_TYPE` with the value of `WP_ENV` for WordPress 5.5 support.
 * Support Composer 2.0.
+* Delete `wp-config.php` after WordPress package is removed.
 
 ### Changed
 * Change default path for `.env` file to only search next to `wp-config.php`. Use `wp-config-env-paths` to change the path(s).